update Channel API

+++ src/controllers/ChannelController.ts
@@ -17,6 +17,16 @@ class ChannelController {
 
 
     public async deleteChannel(req: Request, res: Response) {
+        const channelId : string = req.params.id;
+        const serverId : string = req.query.serverId as string;
+        if(!channelId){
+            res.status(400).json({message : "Channel id is required"});
+            return;
+        }
+        if(!serverId){
+            res.status(400).json({message : "Server id is required"});
+            return;
+        }
         try {
             const deletedChannel = await ChannelService.deleteChannel(req);
             res.json(deletedChannel);
@@ -27,6 +37,16 @@ class ChannelController {
     }
 
     public async updateChannel(req: Request, res: Response) {
+        const channelId : string = req.params.id;
+        const serverId : string = req.query.serverId as string;
+        if(!channelId){
+            res.status(400).json({message : "Channel id is required"});
+            return;
+        }
+        if(!serverId){
+            res.status(400).json({message : "Server id is required"});
+            return;
+        }
         try {
             const channel = await ChannelService.updateChannel(req,res);
             res.json(channel);
